requirements.txt
streamlit
pandas
plotly
openpyxl
folium
shapely
streamlit-folium
xlsxwriter
pillow